Move ArrayBackedPosIntSet and streaming_compression sources into the clp namespace.
