+++ components/core/src/clp/DictionaryReader.hpp
@@ -8,10 +8,10 @@
 #include <string_utils/string_utils.hpp>
 
 #include "../FileReader.hpp"
-#include "../streaming_compression/passthrough/Decompressor.hpp"
-#include "../streaming_compression/zstd/Decompressor.hpp"
 #include "dictionary_utils.hpp"
 #include "DictionaryEntry.hpp"
+#include "streaming_compression/passthrough/Decompressor.hpp"
+#include "streaming_compression/zstd/Decompressor.hpp"
 #include "Utils.hpp"
 
 namespace clp {
