+++ components/core/src/clp/LogTypeDictionaryEntry.hpp
@@ -7,11 +7,11 @@
 #include "../ErrorCode.hpp"
 #include "../FileReader.hpp"
 #include "../ir/types.hpp"
-#include "../streaming_compression/zstd/Compressor.hpp"
-#include "../streaming_compression/zstd/Decompressor.hpp"
 #include "../TraceableException.hpp"
 #include "../type_utils.hpp"
 #include "DictionaryEntry.hpp"
+#include "streaming_compression/zstd/Compressor.hpp"
+#include "streaming_compression/zstd/Decompressor.hpp"
 
 namespace clp {
 /**
