+++ components/core/src/clp/clo/CMakeLists.txt
@@ -37,6 +37,17 @@ set(
         ../streaming_archive/writer/File.hpp
         ../streaming_archive/writer/Segment.cpp
         ../streaming_archive/writer/Segment.hpp
+        ../streaming_compression/Constants.hpp
+        ../streaming_compression/Decompressor.hpp
+        ../streaming_compression/passthrough/Compressor.cpp
+        ../streaming_compression/passthrough/Compressor.hpp
+        ../streaming_compression/passthrough/Decompressor.cpp
+        ../streaming_compression/passthrough/Decompressor.hpp
+        ../streaming_compression/zstd/Compressor.cpp
+        ../streaming_compression/zstd/Compressor.hpp
+        ../streaming_compression/zstd/Constants.hpp
+        ../streaming_compression/zstd/Decompressor.cpp
+        ../streaming_compression/zstd/Decompressor.hpp
         ../Utils.cpp
         ../Utils.hpp
         ../VariableDictionaryEntry.cpp
@@ -80,17 +91,6 @@ set(
         "${PROJECT_SOURCE_DIR}/src/SQLitePreparedStatement.hpp"
         "${PROJECT_SOURCE_DIR}/src/Stopwatch.cpp"
         "${PROJECT_SOURCE_DIR}/src/Stopwatch.hpp"
-        "${PROJECT_SOURCE_DIR}/src/streaming_compression/Constants.hpp"
-        "${PROJECT_SOURCE_DIR}/src/streaming_compression/Decompressor.hpp"
-        "${PROJECT_SOURCE_DIR}/src/streaming_compression/passthrough/Compressor.cpp"
-        "${PROJECT_SOURCE_DIR}/src/streaming_compression/passthrough/Compressor.hpp"
-        "${PROJECT_SOURCE_DIR}/src/streaming_compression/passthrough/Decompressor.cpp"
-        "${PROJECT_SOURCE_DIR}/src/streaming_compression/passthrough/Decompressor.hpp"
-        "${PROJECT_SOURCE_DIR}/src/streaming_compression/zstd/Compressor.cpp"
-        "${PROJECT_SOURCE_DIR}/src/streaming_compression/zstd/Compressor.hpp"
-        "${PROJECT_SOURCE_DIR}/src/streaming_compression/zstd/Constants.hpp"
-        "${PROJECT_SOURCE_DIR}/src/streaming_compression/zstd/Decompressor.cpp"
-        "${PROJECT_SOURCE_DIR}/src/streaming_compression/zstd/Decompressor.hpp"
         "${PROJECT_SOURCE_DIR}/src/StringReader.cpp"
         "${PROJECT_SOURCE_DIR}/src/StringReader.hpp"
         "${PROJECT_SOURCE_DIR}/src/Thread.cpp"
