+++ components/core/src/clp/dictionary_utils.hpp
@@ -4,7 +4,7 @@
 #include <string>
 
 #include "../FileReader.hpp"
-#include "../streaming_compression/Decompressor.hpp"
+#include "streaming_compression/Decompressor.hpp"
 
 namespace clp {
 void open_dictionary_for_reading(
