+++ components/core/src/clp/streaming_archive/writer/Archive.hpp
@@ -14,9 +14,9 @@
 #include <log_surgeon/LogEvent.hpp>
 #include <log_surgeon/ReaderParser.hpp>
 
-#include "../../../ArrayBackedPosIntSet.hpp"
 #include "../../../ErrorCode.hpp"
 #include "../../../ir/LogEvent.hpp"
+#include "../../ArrayBackedPosIntSet.hpp"
 #include "../../GlobalMetadataDB.hpp"
 #include "../../LogTypeDictionaryWriter.hpp"
 #include "../../VariableDictionaryWriter.hpp"
