+++ components/core/src/clp/streaming_archive/writer/Segment.hpp
@@ -6,9 +6,9 @@
 
 #include "../../../Defs.h"
 #include "../../../ErrorCode.hpp"
-#include "../../../streaming_compression/passthrough/Compressor.hpp"
-#include "../../../streaming_compression/zstd/Compressor.hpp"
 #include "../../../TraceableException.hpp"
+#include "../../streaming_compression/passthrough/Compressor.hpp"
+#include "../../streaming_compression/zstd/Compressor.hpp"
 #include "../Constants.hpp"
 
 namespace clp::streaming_archive::writer {
