+++ components/core/src/clp/streaming_compression/passthrough/Decompressor.hpp
@@ -1,18 +1,15 @@
-#ifndef STREAMING_COMPRESSION_ZSTD_DECOMPRESSOR_HPP
-#define STREAMING_COMPRESSION_ZSTD_DECOMPRESSOR_HPP
+#ifndef CLP_STREAMING_COMPRESSION_PASSTHROUGH_DECOMPRESSOR_HPP
+#define CLP_STREAMING_COMPRESSION_PASSTHROUGH_DECOMPRESSOR_HPP
 
-#include <memory>
-#include <string>
-
-#include <boost/iostreams/device/mapped_file.hpp>
-#include <zstd.h>
-
-#include "../../FileReader.hpp"
-#include "../../TraceableException.hpp"
+#include "../../../FileReader.hpp"
+#include "../../../TraceableException.hpp"
 #include "../Decompressor.hpp"
 
-namespace streaming_compression::zstd {
-class Decompressor : public ::streaming_compression::Decompressor {
+namespace clp::streaming_compression::passthrough {
+/**
+ * Decompressor that passes all data through without any decompression.
+ */
+class Decompressor : public ::clp::streaming_compression::Decompressor {
 public:
     // Types
     class OperationFailed : public TraceableException {
@@ -23,19 +20,20 @@ public:
 
         // Methods
         char const* what() const noexcept override {
-            return "streaming_compression::zstd::Decompressor operation failed";
+            return "streaming_compression::passthrough::Decompressor operation failed";
         }
     };
 
-    // Constructor
-    /**
-     * @throw Decompressor::OperationFailed if zstd decompressor stream
-     * cannot be initialized
-     */
-    Decompressor();
+    // Constructors
+    Decompressor()
+            : ::clp::streaming_compression::Decompressor(CompressorType::Passthrough),
+              m_input_type(InputType::NotInitialized),
+              m_compressed_data_buf(nullptr),
+              m_compressed_data_buf_len(0),
+              m_decompressed_stream_pos(0) {}
 
     // Destructor
-    ~Decompressor();
+    ~Decompressor() = default;
 
     // Explicitly disable copy and move constructor/assignment
     Decompressor(Decompressor const&) = delete;
@@ -47,11 +45,9 @@ public:
      * @param buf
      * @param num_bytes_to_read The number of bytes to try and read
      * @param num_bytes_read The actual number of bytes read
-     * @return Same as FileReader::try_read if the decompressor is attached to a file
      * @return ErrorCode_NotInit if the decompressor is not open
      * @return ErrorCode_BadParam if buf is invalid
      * @return ErrorCode_EndOfFile on EOF
-     * @return ErrorCode_Failure on decompression failure
      * @return ErrorCode_Success on success
      */
     ErrorCode try_read(char* buf, size_t num_bytes_to_read, size_t& num_bytes_read) override;
@@ -59,7 +55,7 @@ public:
      * Tries to seek from the beginning to the given position
      * @param pos
      * @return ErrorCode_NotInit if the decompressor is not open
-     * @return Same as ReaderInterface::try_read_exact_length
+     * @return ErrorCode_Truncated if the position is past the last byte in the file
      * @return ErrorCode_Success on success
      */
     ErrorCode try_seek_from_begin(size_t pos) override;
@@ -81,7 +77,7 @@ public:
      * @param decompressed_stream_pos
      * @param extraction_buf
      * @param extraction_len
-     * @return Same as streaming_compression::zstd::Decompressor::try_seek_from_begin
+     * @return Same as streaming_compression::passthrough::Decompressor::try_seek_from_begin
      * @return Same as ReaderInterface::try_read_exact_length
      */
     ErrorCode get_decompressed_stream_region(
@@ -90,53 +86,22 @@ public:
             size_t extraction_len
     ) override;
 
-    // Methods
-    /***
-     * Initialize streaming decompressor to decompress from a compressed file specified by the
-     * given path
-     * @param compressed_file_path
-     * @param decompressed_stream_block_size
-     * @return ErrorCode_Failure if the provided path cannot be memory mapped
-     * @return ErrorCode_Success on success
-     */
-    ErrorCode open(std::string const& compressed_file_path);
-
 private:
-    // Enum class
     enum class InputType {
-        // Note: do nothing but generate an error to prevent this required
-        // parameter is not initialized properly
         NotInitialized,
         CompressedDataBuf,
-        MemoryMappedCompressedFile,
         File
     };
-
-    // Methods
-    /**
-     * Reset streaming decompression state so it will start decompressing from the beginning of
-     * the stream afterwards
-     */
-    void reset_stream();
 
     // Variables
     InputType m_input_type;
 
-    // Compressed stream variables
-    ZSTD_DStream* m_decompression_stream;
-
-    boost::iostreams::mapped_file_source m_memory_mapped_compressed_file;
     FileReader* m_file_reader;
-    size_t m_file_reader_initial_pos;
-    std::unique_ptr<char[]> m_file_read_buffer;
-    size_t m_file_read_buffer_length;
-    size_t m_file_read_buffer_capacity;
-
-    ZSTD_inBuffer m_compressed_stream_block;
+    char const* m_compressed_data_buf;
+    size_t m_compressed_data_buf_len;
 
     size_t m_decompressed_stream_pos;
-    size_t m_unused_decompressed_stream_block_size;
-    std::unique_ptr<char[]> m_unused_decompressed_stream_block_buffer;
 };
-}  // namespace streaming_compression::zstd
-#endif  // STREAMING_COMPRESSION_ZSTD_DECOMPRESSOR_HPP
+}  // namespace clp::streaming_compression::passthrough
+
+#endif  // CLP_STREAMING_COMPRESSION_PASSTHROUGH_DECOMPRESSOR_HPP
