+++ components/core/src/clp/streaming_compression/zstd/Constants.hpp
@@ -1,0 +1,11 @@
+#ifndef CLP_STREAMING_COMPRESSION_ZSTD_CONSTANTS_HPP
+#define CLP_STREAMING_COMPRESSION_ZSTD_CONSTANTS_HPP
+
+#include <cstddef>
+#include <cstdint>
+
+namespace clp::streaming_compression::zstd {
+constexpr int cDefaultCompressionLevel = 3;
+}  // namespace clp::streaming_compression::zstd
+
+#endif  // CLP_STREAMING_COMPRESSION_ZSTD_CONSTANTS_HPP
