--- components/core/src/streaming_compression/zstd/Constants.hpp
@@ -1,11 +1,0 @@
-#ifndef STREAMING_COMPRESSION_ZSTD_CONSTANTS_HPP
-#define STREAMING_COMPRESSION_ZSTD_CONSTANTS_HPP
-
-#include <cstddef>
-#include <cstdint>
-
-namespace streaming_compression::zstd {
-constexpr int cDefaultCompressionLevel = 3;
-}  // namespace streaming_compression::zstd
-
-#endif  // STREAMING_COMPRESSION_ZSTD_CONSTANTS_HPP
